Faz o deploy da app de dominio no docker (#18)
* adiciona o deploy da app de dominio no docker

* ajuste no comando --run

+++ Platform.Cli/actions/domain/cleanAppAction.js
@@ -3,22 +3,31 @@
  * @description Esta action deleta o pacote de execução da aplicação
  */
 
-var AppInstance = require("../../app_instance");
+const AppInstance = require("../../app_instance");
+const DockerService = require("../../services/docker");
 module.exports = class CleanAppAction{
     constructor(){
         this.appInstance = new AppInstance();
+        this.docker = new DockerService();
     }
 
     /**
-     * 
+     *
      * @param {String} appId id da plataforma gerado pelo cli
      * @description apaga a pasta da aplicação no diretorio temporario
      */
-    clean(){        
+    clean(){
         var config = this.appInstance.getLockInstance();
+        var env = this.appInstance.getAppConfig();
         var shell = require("shelljs");
         var os = require("os");
-        shell.rm("-rf",os.tmpdir()+"/"+config.id);
-        shell.rm("./plataforma.instance.lock");
+        var fs = require("fs");
+        if (fs.existsSync(os.tmpdir()+"/"+config.id)){
+            shell.rm("-rf",os.tmpdir()+"/"+config.id);
+        }
+        if (fs.existsSync("./plataforma.instance.lock")){
+            shell.rm("./plataforma.instance.lock");
+        }
+        this.docker.rm({conf:env});
     }
-}
+};
